refactor(shelter): extract symmetricBoundsAroundPinCentroid into Map/utils
Moves the pin-fit bounds helper out of useShelterMapSearch into
Map/utils alongside the other bounds helpers, with unit tests covering
the centroid-centered bounds and the minimum-padding behavior. This was
the last piece of untested pure logic in the feature.
