+++ libs/react/shelter/src/lib/components/Map/utils/symmetricBoundsAroundPinCentroid.ts
@@ -1,0 +1,48 @@
+import {
+  DEFAULT_BOUNDS_MILES,
+  MILES_TO_DEGREES_AT_EQUATOR,
+} from '../constants.maps';
+import { TLatLng } from '../types.maps';
+
+/**
+ * Builds a LatLngBounds symmetric around the centroid of pins so `fitBounds`
+ * centers the map on that centroid (not on a corner of an asymmetric cluster)
+ * while keeping every pin inside the bounds, with a minimum padding radius.
+ */
+export function symmetricBoundsAroundPinCentroid(
+  pinLocations: TLatLng[]
+): google.maps.LatLngBounds {
+  const n = pinLocations.length;
+  const centroidLat = pinLocations.reduce((sum, p) => sum + p.latitude, 0) / n;
+  const centroidLng = pinLocations.reduce((sum, p) => sum + p.longitude, 0) / n;
+
+  let maxHalfLatDeg = 0;
+  let maxHalfLngDeg = 0;
+
+  for (const p of pinLocations) {
+    maxHalfLatDeg = Math.max(maxHalfLatDeg, Math.abs(p.latitude - centroidLat));
+    maxHalfLngDeg = Math.max(
+      maxHalfLngDeg,
+      Math.abs(p.longitude - centroidLng)
+    );
+  }
+
+  const minHalfLatDeg = DEFAULT_BOUNDS_MILES / 2 / MILES_TO_DEGREES_AT_EQUATOR;
+  const cosLat = Math.cos((centroidLat * Math.PI) / 180) || 1e-6;
+  const minHalfLngDeg =
+    DEFAULT_BOUNDS_MILES / 2 / (MILES_TO_DEGREES_AT_EQUATOR * cosLat);
+
+  const halfLat = Math.max(maxHalfLatDeg, minHalfLatDeg);
+  const halfLng = Math.max(maxHalfLngDeg, minHalfLngDeg);
+
+  const sw = {
+    lat: centroidLat - halfLat,
+    lng: centroidLng - halfLng,
+  };
+  const ne = {
+    lat: centroidLat + halfLat,
+    lng: centroidLng + halfLng,
+  };
+
+  return new google.maps.LatLngBounds(sw, ne);
+}
